Renames simulationDataIO file
